requirements.txt
viser
numpy
scipy
rich
hydra-core
omegaconf
